feat(commons): 🎸 Change LEAF-Writer instantiation. Improve topbar, saving feedback, and main menu
LEAF-Writer must be instantiate just with the container element. TO
render the editor, we need to run the method `init` passing the document
and other options. Top bar has improve UX to show the document full path
and saving feedback. Main Menu was redesigns and has new options: 1. new
from a template, 2. Open Recent.

BREAKING CHANGE: 🧨 LEAF-Writer must be instantiate just with the container element. TO
render the editor, we need to run the method `init` passing the document
and other options.

✅ Closes: #39