feat(storage service): 🎸 Capture http error 409 (conflict) while saving
This error is generated by GitHub (and possibly by GitLab) when the
request for saving (overwriting) a file does not have a matching hash.
It also caused by a delay on GitHub side to update the hash of the
document, which results on not saving the file. This commit also add
translations and adjusts tests accordingly.

BREAKING CHANGE: 🧨 As a consequence of this change, results form the headless modulo
changed a bit. The object returned there is an error changed from
`{error: string}` to `{type: string; message: string}`