feat(commons): 🎸 Introduce autosave
LEAF-Writer will set a 30s timer every time the document becomes
'modified'. After this time, LW will trigger the save function. Users
can also trigger the save funciotion, which, in this case will take
priority ove the timer. The ttimer resets  and stops after the  file is
saved. If a the provider respond with a conflict (error 409), LW will
make 5 subsquent attempts to save the file in a 10s interval. If the
error persists, it continues to try to save the file after every  30s.