feat(commons): 🎸 Enable not registered users to use LEAF-Writer
Complete redeisgn of the homepage with overhauled storage panel. Not
registered users can open templates and sample documents as well as
upload and paste xml from their devices. However, they are not allow to
save to the cloud. An anonymous user is created and used on web
annotations. User can sign in from the editor view.