feat(validator): 🎸 Refactoring conversion and caching. LW-Validator now saves cache on the brower indexedDB
This fix a mistake where the hash of the file was not properly compared
with the incoming file. The schema is now cached on the indexedDBm which
there is much more space than the localStorage. As such, we do not need
to compressed salve's processed file.  IndexedDB is available to
websorker (localstorage is not). Thus, the validator will not pass the
processed schema back to LEAF-Writer. It will preapare, store, and
retrieve the cached schema itselfe.

BREAKING CHANGE: 🧨 Changes the 'initialize' parameters and response. Check documentation
for more info.